cmd/branch-compare: move logic to branch_diff module
The idea here is to reuse this for other commands especially
those that involve running scripts against different branches
like `brew service-diff`.
